delegator set api returning top
